splitted TransformWithCovariance
