+++ src/CMakeLists.txt
@@ -1,7 +1,18 @@
-set(SOURCES Angle.cpp Pose.cpp TwistWithCovariance.cpp Waypoint.cpp)
+set(SOURCES
+        Angle.cpp
+        Pose.cpp
+        TransformWithCovariance.cpp
+        TwistWithCovariance.cpp
+        Waypoint.cpp
+)
 
-set(HEADERS Logging.hpp Singleton.hpp logging/logging_iostream_style.h logging/logging_printf_style.h
-            Wrench.hpp)
+set(HEADERS 
+        Logging.hpp 
+        Singleton.hpp 
+        logging/logging_iostream_style.h 
+        logging/logging_printf_style.h
+        Wrench.hpp
+)
 
 
 # Using SISL as optional dependency
